コンポーネントを分割
